+++ Functions/ConvertFrom-Base64ToPem.ps1
@@ -1,0 +1,11 @@
+function ConvertFrom-Base64ToPem {
+        
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory=$True)]
+        [ValidateNotNullOrEmpty()]
+        [string]$String
+    )
+
+    return "-----BEGIN CERTIFICATE-----`n$(($String -split "(.{64})" | Where-Object {$_ -match '\S'}) -join "`n")`n-----END CERTIFICATE-----"
+}