+++ Functions/ConvertFrom-PemToBase64.ps1
@@ -1,0 +1,11 @@
+function ConvertFrom-PemToBase64 {
+
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory=$True)]
+        [ValidateNotNullOrEmpty()]
+        [string]$String
+    )
+
+    return $String -replace "-----.*-----", "" -replace "\r", "" -replace "\n", ""
+}