+++ Functions/Get-ESTCACertificates.ps1
@@ -20,7 +20,7 @@ Function Get-ESTCACertificates {
     begin {
         
         # This hides the Status Indicators of the Invoke-WebRequest Calls later on
-        $ProgressPreference = 'SilentlyContinue'
+        $ProgressPreference = "SilentlyContinue"
     }
 
     process {