+++ Functions/Get-WSTEPResponse.ps1
@@ -38,7 +38,7 @@ Function Get-WSTEPResponse {
     begin {
 
         # This hides the Status Indicators of the Invoke-WebRequest Calls later on
-        $ProgressPreference = 'SilentlyContinue'
+        $ProgressPreference = "SilentlyContinue"
 
     }