+++ Functions/Get-XCEPEnrollmentPolicy.ps1
@@ -24,7 +24,7 @@ Function Get-XCEPEnrollmentPolicy {
     begin {
 
         # This hides the Status Indicators of the Invoke-WebRequest Calls later on
-        $ProgressPreference = 'SilentlyContinue'
+        $ProgressPreference = "SilentlyContinue"
 
     }